+++ tests/from_to_sql.rs
@@ -487,10 +487,10 @@ fn raw_from_to_sql() -> Result<()> {
 #[test]
 fn timestamp_from_sql() -> Result<()> {
     let conn = common::connect()?;
-    let ts = Timestamp::new(2012, 3, 4, 0, 0, 0, 0);
+    let ts = Timestamp::new(2012, 3, 4, 0, 0, 0, 0)?;
 
     test_from_sql!(&conn, "DATE '2012-03-04'", &OracleType::Date, &ts);
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?;
     test_from_sql!(
         &conn,
         "TO_DATE('2012-03-04 05:06:07', 'YYYY-MM-DD HH24:MI:SS')",
@@ -504,21 +504,21 @@ fn timestamp_from_sql() -> Result<()> {
         &OracleType::Timestamp(0),
         &ts
     );
-    let ts = ts.and_prec(1);
+    let ts = ts.and_prec(1)?;
     test_from_sql!(
         &conn,
         "CAST(TO_DATE('2012-03-04 05:06:07', 'YYYY-MM-DD HH24:MI:SS') AS TIMESTAMP(1))",
         &OracleType::Timestamp(1),
         &ts
     );
-    let ts = ts.and_prec(6);
+    let ts = ts.and_prec(6)?;
     test_from_sql!(
         &conn,
         "CAST(TO_DATE('2012-03-04 05:06:07', 'YYYY-MM-DD HH24:MI:SS') AS TIMESTAMP)",
         &OracleType::Timestamp(6),
         &ts
     );
-    let ts = ts.and_prec(9);
+    let ts = ts.and_prec(9)?;
     test_from_sql!(
         &conn,
         "CAST(TO_DATE('2012-03-04 05:06:07', 'YYYY-MM-DD HH24:MI:SS') AS TIMESTAMP(9))",
@@ -531,21 +531,21 @@ fn timestamp_from_sql() -> Result<()> {
         &OracleType::Timestamp(9),
         &ts
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456789);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456789)?;
     test_from_sql!(
         &conn,
         "TO_TIMESTAMP('2012-03-04 05:06:07.123456789', 'YYYY-MM-DD HH24:MI:SS.FF')",
         &OracleType::Timestamp(9),
         &ts
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456000);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456000)?;
     test_from_sql!(
         &conn,
         "TO_TIMESTAMP('2012-03-04 05:06:07.123456', 'YYYY-MM-DD HH24:MI:SS.FF')",
         &OracleType::Timestamp(9),
         &ts
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123000000);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123000000)?;
     test_from_sql!(
         &conn,
         "TO_TIMESTAMP('2012-03-04 05:06:07.123', 'YYYY-MM-DD HH24:MI:SS.FF')",
@@ -553,21 +553,21 @@ fn timestamp_from_sql() -> Result<()> {
         &ts
     );
 
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0).and_tz_offset(0);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?.and_tz_offset(0)?;
     test_from_sql!(
         &conn,
         "TO_TIMESTAMP_TZ('2012-03-04 05:06:07 +00:00', 'YYYY-MM-DD HH24:MI:SS TZH:TZM')",
         &OracleType::TimestampTZ(9),
         &ts
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0).and_tz_hm_offset(8, 45);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?.and_tz_hm_offset(8, 45)?;
     test_from_sql!(
         &conn,
         "TO_TIMESTAMP_TZ('2012-03-04 05:06:07 +08:45', 'YYYY-MM-DD HH24:MI:SS TZH:TZM')",
         &OracleType::TimestampTZ(9),
         &ts
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0).and_tz_hm_offset(-8, -45);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?.and_tz_hm_offset(-8, -45)?;
     test_from_sql!(
         &conn,
         "TO_TIMESTAMP_TZ('2012-03-04 05:06:07 -08:45', 'YYYY-MM-DD HH24:MI:SS TZH:TZM')",
@@ -580,7 +580,7 @@ fn timestamp_from_sql() -> Result<()> {
 #[test]
 fn timestamp_to_sql() -> Result<()> {
     let conn = common::connect()?;
-    let ts = Timestamp::new(2012, 3, 4, 0, 0, 0, 0);
+    let ts = Timestamp::new(2012, 3, 4, 0, 0, 0, 0)?;
 
     test_to_sql!(
         &conn,
@@ -589,7 +589,7 @@ fn timestamp_to_sql() -> Result<()> {
         "2012-03-04 00:00:00"
     );
 
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?;
     test_to_sql!(
         &conn,
         &ts,
@@ -603,21 +603,21 @@ fn timestamp_to_sql() -> Result<()> {
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS')",
         "2012-03-04 05:06:07"
     );
-    let ts = ts.and_prec(1);
+    let ts = ts.and_prec(1)?;
     test_to_sql!(
         &conn,
         &ts,
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS')",
         "2012-03-04 05:06:07"
     );
-    let ts = ts.and_prec(6);
+    let ts = ts.and_prec(6)?;
     test_to_sql!(
         &conn,
         &ts,
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS')",
         "2012-03-04 05:06:07"
     );
-    let ts = ts.and_prec(9);
+    let ts = ts.and_prec(9)?;
     test_to_sql!(
         &conn,
         &ts,
@@ -630,21 +630,21 @@ fn timestamp_to_sql() -> Result<()> {
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS')",
         "2012-03-04 05:06:07"
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456789);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456789)?;
     test_to_sql!(
         &conn,
         &ts,
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS.FF')",
         "2012-03-04 05:06:07.123456789"
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456000);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123456000)?;
     test_to_sql!(
         &conn,
         &ts,
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS.FF6')",
         "2012-03-04 05:06:07.123456"
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123000000);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 123000000)?;
     test_to_sql!(
         &conn,
         &ts,
@@ -652,21 +652,21 @@ fn timestamp_to_sql() -> Result<()> {
         "2012-03-04 05:06:07.123"
     );
 
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0).and_tz_offset(0);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?.and_tz_offset(0)?;
     test_to_sql!(
         &conn,
         &ts,
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS TZH:TZM')",
         "2012-03-04 05:06:07 +00:00"
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0).and_tz_hm_offset(8, 45);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?.and_tz_hm_offset(8, 45)?;
     test_to_sql!(
         &conn,
         &ts,
         "TO_CHAR(:1, 'YYYY-MM-DD HH24:MI:SS TZH:TZM')",
         "2012-03-04 05:06:07 +08:45"
     );
-    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0).and_tz_hm_offset(-8, -45);
+    let ts = Timestamp::new(2012, 3, 4, 5, 6, 7, 0)?.and_tz_hm_offset(-8, -45)?;
     test_to_sql!(
         &conn,
         &ts,